chore(group_create_tag): 태그 입력 폼을 컴포넌트로 작성
재사용성을 위해 컴포넌트화 시킴
